chore: remove redundant code lines
