--- src/utils/DB/errors/NoRequireEntity.error.ts
@@ -1,6 +1,0 @@
-export class NoRequiredEntity extends Error {
-  constructor(operation: string) {
-    super(`Fail during ${operation}.`);
-    this.name = 'No required entity';
-  }
-}
