+++ src/utils/database/entities/DBAlbums.ts
@@ -9,17 +9,6 @@ export default class DBAlbums extends DBEntity<
   UpdateAlbumDto,
   CreateAlbumDto
 > {
-  // constructor() {
-  //   super();
-  //   this.entities = [
-  //     {
-  //       id: 'b018b96f-0f27-4327-8e06-ca0e4be11c37',
-  //       name: 'name',
-  //       year: 2000,
-  //       artistId: 'b018b96f-0f27-4327-8e06-ca0e4be11c37',
-  //     },
-  //   ];
-  // }
   async create(dto: CreateAlbumDto) {
     const created: Album = {
       id: uuid.v4(),
